HParser.FirstSet

Description

Set of functions to compute first sets in a Grammar.

Documentation

ruleFirstSet :: Grammar -> Rule -> Set String

Computes the first set of a rule in a grammar. If the resulting set contains the empty string (""), epsilon is in the set.

Returns a set of strings, the names of the terminals in the first set.

>>> grammar
S 	-> E Sp
Sp 	-> ε
Sp 	-> 'PLUS' S
E 	-> 'ONE'
E 	-> 'OPEN' S 'CLOSE'
>>> rule
S 	-> E Sp
>>> ruleFirstSet grammar rule
fromList ["ONE","OPEN"]

symbolFirstSet :: Grammar -> Symbol -> Set String

Computes the first set of a symbol in a grammar. If the resulting set contains the empty string (""), epsilon is in the set. This includes terminals, which have a trivial solution (singleton set of itself).

Returns a set of strings, the names of the terminals in the first set.

>>> grammar
S 	-> E Sp
Sp 	-> ε
Sp 	-> 'PLUS' S
E 	-> 'ONE'
E 	-> 'OPEN' S 'CLOSE'
>>> symbolFirstSet grammar (NonTerminal Sp)
fromList ["","PLUS"]
>>> symbolFirstSet grammar (Terminal "ONE")
fromList ["ONE"]

firstSets :: Grammar -> [Set String]

For testing and debugging purposes. Combination function for the first sets of all rules in the grammar. If the resulting set contains the empty string (""), epsilon is in the set.

Returns a list of sets of strings, one set per rule in the grammar, in the same order.

>>> grammar
S 	-> E Sp
Sp 	-> ε
Sp 	-> 'PLUS' S
E 	-> 'ONE'
E 	-> 'OPEN' S 'CLOSE'
>>> firstSets grammar
[fromList ["ONE","OPEN"],fromList [""],fromList ["PLUS"],fromList ["ONE"],fromList ["OPEN"]]