HParser.FollowSet

Description

Set of functions to compute follow sets in a Grammar.

Documentation

followSet :: Grammar -> Symbol -> Set String

Computes the follow set of a given non-terminal in a grammar. If the resulting set contains the empty string (""), $, or end-of-string, is in the set.

>>> grammar
S 	-> E Sp
Sp 	-> ε
Sp 	-> 'PLUS' S
E 	-> 'ONE'
E 	-> 'OPEN' S 'CLOSE'
>>> followSet grammar (NonTerminal "E")
fromList ["","CLOSE","PLUS"]