HParser.Generator

Description

Set of functions to generate LL(1) parsers from a Grammar.

Documentation

genParser :: Grammar -> String -> String

Creates a Haskell LL(1)-parser from given Grammar, with a given name. It will give warnings or errors if the grammar is not well-defined or not LL(1).

Example:

>>> grammar
S 	-> E Sp
Sp 	-> ε
Sp 	-> PLUS S
E 	-> ONE
E 	-> OPEN S CLOSE
>>> putStr $ genParser "Parser" grammar
module Parser (Token (..), TokenTuple (..), ParseTree (..), parser, parseTree, printParseTree) where

import Data.Tree
import Debug.Trace
import Control.Arrow

-- GRAMMAR-SPECIFIC PARSER CODE
data Token = PLUS | ONE | OPEN | CLOSE
   deriving (Read, Show, Eq)

data NonTerminal = S | Sp | E
   deriving (Read, Show, Eq)

instance Symbol NonTerminal where
   parseEOF Sp = True
   parseEOF _ = False

   parseRule S ONE = parse E >>> parse Sp
   parseRule S OPEN = parse E >>> parse Sp
   parseRule Sp CLOSE = parseEpsilon
   parseRule Sp PLUS = parseToken PLUS >>> parse S
   parseRule E ONE = parseToken ONE
   parseRule E OPEN = parseToken OPEN >>> parse S >>> parseToken CLOSE
   parseRule _ _ = parseFailure

-- Set starting symbol
parser = _parser S
parseTree = _parseTree S

(omitting standard parser code)

printParser :: Grammar -> String -> IO ()

Uses the genParser function to print a parser to IO.

saveParser :: Grammar -> String -> IO ()

Uses the genParser function to save a parser to a file. The name that is needed for generating the parser, is also used as the filename.

checkSets :: Grammar -> String

Function to check for conflicting rules.

Returns conflicts written in a string. If a grammar is LL(1), this will return empty string ("").