HParser.Grammar

Description

This module defined the core data types used, specifically grammars.

Documentation

data Symbol

Represents a symbol in a grammar, in other word either a terminal, non-terminal or epsilon.

Symbol is an instance is Show:

>>> Terminal "TOKEN"
'TOKEN'
>>> NonTerminal S
S
>>> Epsilon
ε

Constructors

Terminal String

NonTerminal String

Epsilon

data Rule

Respresents a single rule in a grammar. Note that a non-terminal can be in the left-hand side of multiple rules in a grammar.

Rule is an instance is Show:

>>> Rule (NonTerminal "E") [(Terminal "OPEN"), (NonTerminal "S"), (Terminal "CLOSE")]
E 	-> 'OPEN' S 'CLOSE'

Constructors

Rule Symbol [Symbol]

data Grammar

Represents a single cohesive set of rules, in other words a grammar. For example:

grammar = Grammar (NonTerminal "S") [
   Rule (NonTerminal "S") [(NonTerminal "E"), (NonTerminal "Sp")],
   Rule (NonTerminal "Sp") [],
   Rule (NonTerminal "Sp") [(Terminal "PLUS"), (NonTerminal "S")],
   Rule (NonTerminal "E") [(Terminal "ONE")],
   Rule (NonTerminal "E") [(Terminal "OPEN"), (NonTerminal "S"), (Terminal "CLOSE")]
   ]

Grammar is an instance is Show:

>>> grammar
S 	-> E Sp
Sp 	-> ε
Sp 	-> 'PLUS' S
E 	-> 'ONE'
E 	-> 'OPEN' S 'CLOSE'

Constructors

Grammar Symbol [Rule]

nonTerminalNames :: Grammar -> [String]

Returns all non-terminals in a grammar, as a list of strings (symbol names).

terminalNames :: Grammar -> [String]

Returns all terminals in a grammar, as a list of strings (symbol names).

startSymbol :: Grammar -> String

Returns the starting symbol of a grammar. Returns empty stirng if one is not properly defined.

syntaxSanity :: Grammar -> String

Checks whether a grammar has been properly defined. It does the following checks:

Check whether every rule has a non-terminal left-hand side.

Check whether the starting symbol is a non-terminal.

For example:

grammar = Grammar (Terminal "Token") [
   Rule (NonTerminal "S") [(Terminal "Token")],
   Rule (Terminal "Token") []
   ]

>>> grammar
S 	-> 'Token'
'Token' 	-> ε
>>> putStr $ syntaxSanity grammar
Error: Start symbol must be a non-terminal
The following rule has an invalid left hand side: 
   'Token' 	-> ε